requirements.txt
# PyTorch + torchvision are installed by the Dockerfile with the cu118 index URL.
# A bare version spec here lets pip override that with the latest cu130 build,
# which doesn't work with the host GPU driver (supports up to CUDA 12.x).
#
# For local (non-Docker) installs run this first:
#   pip install "torch>=2.2.0,<2.3.0" torchvision \
#       --index-url https://download.pytorch.org/whl/cu118

# Vision models
ultralytics==8.3.40

# Tabular models
scikit-learn==1.4.2
joblib==1.4.2

# Data processing
numpy==1.26.4
pandas==2.2.2

# ONNX export (Edge Impulse ingestion)
onnx==1.16.0
onnxruntime==1.18.0
skl2onnx==1.17.0

# Image handling (required by ultralytics)
pillow==10.3.0

# Config parsing
pyyaml==6.0.1

# Optional — only needed if re-downloading the Roboflow dataset
# roboflow==1.1.48
